fix: remove unused components
